requirements.txt
# Core dependencies
torch>=2.0.0
pillow>=9.0.0
numpy<2.0.0
gradio>=4.36.0
ultralytics>=8.3.0,<9.0.0
opencv-python-headless==4.10.0.84

# Additional dependencies
scikit-learn>=1.0.0
transformers>=4.30.0

# Optional but recommended for better performance
# torchvision>=0.15.0  # Uncomment if you need torchvision
# onnx>=1.12.0  # Uncomment for ONNX export support
